package, static util, naming
